다시 작성클론 코딩한 쇼핑몰 !

+++ remind-app-shop-1/src/main/java/jpa/app/shop/controller/form/BookForm.java
@@ -8,8 +8,8 @@ public class BookForm {
 
 	private Long id;
 	private String name;
-	private int price;
-	private int stockQuantity;
+	private long price;
+	private long stockQuantity;
 	private String author;
 	private String isbn;
 }
